Backend/app/schemas/auth.py
from __future__ import annotations

from pydantic import BaseModel, EmailStr, Field

from app.db.models.user import UserRole


class UserRegister(BaseModel):
    name: str = Field(min_length=1, max_length=255)
    email: EmailStr
    password: str = Field(min_length=8, max_length=128)


class Token(BaseModel):
    access_token: str
    token_type: str = "bearer"


class TokenPayload(BaseModel):
    sub: EmailStr
    role: UserRole

